TMS/CourseCreation: factored course creation out.
